Refactor shared folder structure
