+++ src/app/blog/blog.component.spec.ts
@@ -8,7 +8,7 @@ import {
 import { of } from 'rxjs';
 
 import { BlogComponent } from './blog.component';
-import { BlogsService } from '../shared/blogs/blogs.service';
+import { BlogsService } from '../shared/services/blogs/blogs.service';
 
 const blogsServiceMock = {
   getAllBlogs: () => of([])
