+++ src/app/get-involved/country-supporters.service.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { AngularFirestore } from '@angular/fire/firestore';
 
-import { DataService } from '../shared/data-service/data.service';
+import { DataService } from '../shared/services/data/data.service';
 import { CountrySupporters } from '../shared/models/country-supporters.model';
-import { FirebaseCollections } from '../shared/data-service/collections';
+import { FirebaseCollections } from '../shared/services/data/collections';
 
 @Injectable()
 export class CountrySupportersService extends DataService<CountrySupporters> {
