+++ src/app/home/activity-indicators/activity-indicators.service.ts
@@ -3,8 +3,8 @@ import { HttpClient } from '@angular/common/http';
 import { AngularFirestore } from '@angular/fire/firestore';
 
 import { ActivityIndicator } from '../../shared/models/activity-indicator.model';
-import { DataService } from '../../shared/data-service/data.service';
-import { FirebaseCollections } from '../../shared/data-service/collections';
+import { DataService } from '../../shared/services/data/data.service';
+import { FirebaseCollections } from '../../shared/services/data/collections';
 
 @Injectable()
 export class ActivityIndicatorsService extends DataService<ActivityIndicator> {
