+++ src/app/home/home.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
-import { BlogsService } from '../shared/blogs/blogs.service';
+import { BlogsService } from '../shared/services/blogs/blogs.service';
 import { SharedModule } from '../shared/shared.module';
 import { HomeComponent } from './home.component';
 import { BannerComponent } from './banner/banner.component';
