+++ src/app/home/top-blogs/top-blogs.component.ts
@@ -3,8 +3,8 @@ import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { Blog } from 'src/app/shared/models/blog.model';
-import { BlogsService } from 'src/app/shared/blogs/blogs.service';
-import { CloudFunctions } from 'src/app/shared/data-service/cloud-functions';
+import { BlogsService } from 'src/app/shared/services/blogs/blogs.service';
+import { CloudFunctions } from 'src/app/shared/services/data/cloud-functions';
 import { DataOrder } from 'src/app/shared/models/data-type.model';
 
 const MAX_TOP_BLOGS_LENGTH = 3;
