+++ src/app/shared/components/waiting-spinner/waiting-spinner.component.scss
